+++ python/metatensor-operations/tests/split.py
@@ -15,20 +15,18 @@ def test_split_block_samples():
     tensor = metatensor.load(os.path.join(DATA_ROOT, "qm7-spherical-expansion.npz"))
     block = tensor.block(o3_lambda=2, center_type=6, neighbor_type=6)
 
-    grouped_labels = [
+    selections = [
         Labels(names=["system"], values=np.array([[0], [6], [7]])),
         Labels(names=["system"], values=np.array([[2], [3], [4]])),
         Labels(names=["system"], values=np.array([[1], [5], [8], [9]])),
     ]
 
-    splitted = metatensor.split_block(
-        block, axis="samples", grouped_labels=grouped_labels
-    )
+    splitted = metatensor.split_block(block, axis="samples", selections=selections)
 
     assert len(splitted) == 3
     assert sum(len(b.samples) for b in splitted) == len(block.samples)
 
-    for split_block, expected_samples in zip(splitted, grouped_labels):
+    for split_block, expected_samples in zip(splitted, selections):
         structure_values = np.unique(split_block.samples["system"]).reshape(-1, 1)
         assert np.all(structure_values == expected_samples.values)
 
@@ -65,17 +63,15 @@ def test_split_block_samples_not_everything():
     tensor = metatensor.load(os.path.join(DATA_ROOT, "qm7-spherical-expansion.npz"))
     block = tensor.block(o3_lambda=2, center_type=6, neighbor_type=6)
 
-    # using `grouped_labels` with some samples not present in the initial block,
+    # using `selections` with some samples not present in the initial block,
     # and not including all samples
-    grouped_labels = [
+    selections = [
         # 1 is in the samples, 12 is not
         Labels(names=["system"], values=np.array([[1], [12]])),
         # 18 is not in the samples
         Labels(names=["system"], values=np.array([[18]])),
     ]
-    splitted = metatensor.split_block(
-        block, axis="samples", grouped_labels=grouped_labels
-    )
+    splitted = metatensor.split_block(block, axis="samples", selections=selections)
 
     assert len(splitted) == 2
     partial = splitted[0]
@@ -88,25 +84,25 @@ def test_split_block_samples_not_everything():
     assert empty.components == block.components
     assert empty.properties == block.properties
 
-    # an empty list of grouped_labels gets you an empty list of blocks
-    assert metatensor.split_block(block, axis="samples", grouped_labels=[]) == []
+    # an empty list of selections gets you an empty list of blocks
+    assert metatensor.split_block(block, axis="samples", selections=[]) == []
 
 
 def test_split_samples():
     tensor = metatensor.load(os.path.join(DATA_ROOT, "qm7-spherical-expansion.npz"))
 
-    grouped_labels = [
+    selections = [
         Labels(names=["system"], values=np.array([[0], [6], [7]])),
         Labels(names=["system"], values=np.array([[2], [3], [4]])),
         Labels(names=["system"], values=np.array([[1], [5], [8], [9]])),
     ]
 
-    splitted = metatensor.split(tensor, axis="samples", grouped_labels=grouped_labels)
+    splitted = metatensor.split(tensor, axis="samples", selections=selections)
 
     for split_tensor in splitted:
         assert split_tensor.keys == tensor.keys
 
-    for split_tensor, expected_samples in zip(splitted, grouped_labels):
+    for split_tensor, expected_samples in zip(splitted, selections):
         for split_block, block in zip(split_tensor, tensor):
             split_structures = Labels(
                 ["system"],
@@ -122,8 +118,8 @@ def test_split_samples():
             assert split_block.components == block.components
             assert split_block.properties == block.properties
 
-    # an empty list of grouped_labels gets you an empty list of tensors
-    assert metatensor.split(tensor, axis="samples", grouped_labels=[]) == []
+    # an empty list of selections gets you an empty list of tensors
+    assert metatensor.split(tensor, axis="samples", selections=[]) == []
 
 
 def test_split_block_properties():
@@ -131,17 +127,15 @@ def test_split_block_properties():
     tensor = metatensor.load(os.path.join(DATA_ROOT, "qm7-power-spectrum.npz"))
     block = tensor.block(center_type=8, neighbor_1_type=6, neighbor_2_type=8)
 
-    grouped_labels = [
+    selections = [
         Labels(names=["l", "n_2"], values=np.array([[0, 0], [1, 3], [3, 1]])),
         Labels(names=["l", "n_2"], values=np.array([[4, 2], [4, 3], [4, 1]])),
         Labels(names=["l", "n_2"], values=np.array([[3, 2], [1, 1]])),
     ]
-    splitted = metatensor.split_block(
-        block, axis="properties", grouped_labels=grouped_labels
-    )
+    splitted = metatensor.split_block(block, axis="properties", selections=selections)
 
     assert len(splitted) == 3
-    for split_block, expected_properties in zip(splitted, grouped_labels):
+    for split_block, expected_properties in zip(splitted, selections):
         # no changes to samples & components
         assert split_block.samples == block.samples
         assert split_block.components == block.components
@@ -166,27 +160,25 @@ def test_split_block_properties():
 
             assert np.all(split_gradient.values == gradient.values[..., mask])
 
-    # an empty list of grouped_labels gets you an empty list of blocks
-    assert metatensor.split_block(block, axis="properties", grouped_labels=[]) == []
+    # an empty list of selections gets you an empty list of blocks
+    assert metatensor.split_block(block, axis="properties", selections=[]) == []
 
 
 def test_split_properties():
     # TensorMap with multiple properties
     tensor = metatensor.load(os.path.join(DATA_ROOT, "qm7-power-spectrum.npz"))
 
-    grouped_labels = [
+    selections = [
         Labels(names=["l", "n_2"], values=np.array([[0, 0], [1, 3], [3, 1]])),
         Labels(names=["l", "n_2"], values=np.array([[4, 2], [4, 3], [4, 1]])),
         Labels(names=["l", "n_2"], values=np.array([[3, 2], [1, 1]])),
     ]
-    splitted = metatensor.split(
-        tensor, axis="properties", grouped_labels=grouped_labels
-    )
+    splitted = metatensor.split(tensor, axis="properties", selections=selections)
 
     for split_tensor in splitted:
         assert split_tensor.keys == tensor.keys
 
-    for split_tensor, expected_properties in zip(splitted, grouped_labels):
+    for split_tensor, expected_properties in zip(splitted, selections):
         for split_block, block in zip(split_tensor, tensor):
             # no changes to samples & components
             assert split_block.samples == block.samples
@@ -195,14 +187,14 @@ def test_split_properties():
             for p in split_block.properties.view(expected_properties.names):
                 assert p in expected_properties
 
-    # an empty list of grouped_labels gets you an empty list of blocks
-    assert metatensor.split(tensor, axis="properties", grouped_labels=[]) == []
+    # an empty list of selections gets you an empty list of blocks
+    assert metatensor.split(tensor, axis="properties", selections=[]) == []
 
 
 def test_split_errors():
     tensor = metatensor.load(os.path.join(DATA_ROOT, "qm7-spherical-expansion.npz"))
     block = tensor.block(4)
-    grouped_labels = [
+    selections = [
         Labels(names=["system"], values=np.array([[0], [6], [7]])),
         Labels(names=["system"], values=np.array([[2], [3], [4]])),
         Labels(names=["system"], values=np.array([[1], [5], [8], [9]])),
@@ -213,40 +205,38 @@ def test_split_errors():
         "not <class 'metatensor.block.TensorBlock'>"
     )
     with pytest.raises(TypeError, match=message):
-        metatensor.split(block, axis="samples", grouped_labels=grouped_labels)
+        metatensor.split(block, axis="samples", selections=selections)
 
     message = "axis must be a string, not <class 'float'>"
     with pytest.raises(TypeError, match=message):
-        metatensor.split(tensor, axis=3.14, grouped_labels=grouped_labels)
+        metatensor.split(tensor, axis=3.14, selections=selections)
 
     message = "axis must be either 'samples' or 'properties'"
     with pytest.raises(ValueError, match=message):
-        metatensor.split(tensor, axis="buongiorno!", grouped_labels=grouped_labels)
+        metatensor.split(tensor, axis="buongiorno!", selections=selections)
+
+    message = "`selections` must be a list, " "not <class 'metatensor.labels.Labels'>"
+    with pytest.raises(TypeError, match=message):
+        metatensor.split(tensor, axis="samples", selections=selections[0])
 
     message = (
-        "`grouped_labels` must be a list, " "not <class 'metatensor.labels.Labels'>"
+        "`selection` must be metatensor Labels, an array or List\\[int\\], "
+        "not <class 'str'>"
     )
     with pytest.raises(TypeError, match=message):
-        metatensor.split(tensor, axis="samples", grouped_labels=grouped_labels[0])
+        metatensor.split(tensor, axis="samples", selections=["a", "b", "c"])
 
-    message = "`grouped_labels` elements must be metatensor Labels, not <class 'str'>"
-    with pytest.raises(TypeError, match=message):
-        metatensor.split(tensor, axis="samples", grouped_labels=["a", "b", "c"])
-
-    grouped_labels = [
+    selections = [
         Labels(names=["red"], values=np.array([[0], [6], [7]])),
         Labels(names=["red"], values=np.array([[2], [3], [4]])),
         Labels(names=["wine"], values=np.array([[1], [5], [8], [9]])),
     ]
 
-    message = (
-        "the dimensions names of all Labels in `grouped_labels` must be the same, "
-        "got \\['red'\\] and \\['wine'\\]"
-    )
+    message = "invalid sample name 'red' which is not part of the input"
     with pytest.raises(ValueError, match=message):
-        metatensor.split(tensor, axis="samples", grouped_labels=grouped_labels)
+        metatensor.split(tensor, axis="samples", selections=selections)
 
-    grouped_labels = [
+    selections = [
         Labels(names=["missing", "atom"], values=np.array([[0, 1], [6, 7], [7, 4]])),
         Labels(names=["missing", "atom"], values=np.array([[2, 4], [3, 3], [4, 7]])),
         Labels(
@@ -255,16 +245,13 @@ def test_split_errors():
         ),
     ]
 
-    message = (
-        "the 'missing' dimension name in `grouped_labels` is not part of "
-        "the samples names of the input tensor"
-    )
+    message = "invalid sample name 'missing' which is not part of the input"
     with pytest.raises(ValueError, match=message):
-        metatensor.split(tensor, axis="samples", grouped_labels=grouped_labels)
+        metatensor.split(tensor, axis="samples", selections=selections)
 
     message = (
         "`block` must be a metatensor TensorBlock, "
         "not <class 'metatensor.tensor.TensorMap'>"
     )
     with pytest.raises(TypeError, match=message):
-        metatensor.split_block(tensor, axis="samples", grouped_labels=[])
+        metatensor.split_block(tensor, axis="samples", selections=[])